backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.35
alembic>=1.13.3
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
playwright>=1.47.0
beautifulsoup4>=4.12.3
celery>=5.4.0
redis>=5.1.0
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.9
openai>=1.50.0
langchain>=0.3.0
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4
